Rename: move 'header.utils.js', 'packet-parser.utils.js' to src/utils/packet directory Add: packet-encoder.utils.js to encode packets with proto message type
